+++ .travis.yml
@@ -6,6 +6,7 @@ language: go
 
 go:
     - 1.2.2
+    - 1.3
     - tip
 
 matrix:
@@ -18,12 +19,13 @@ before_install:
     - go get -v github.com/mattn/goveralls
 
 install:
-    - go get -d -v ./... && go build -v .
+    - go get -d -v ./... && go build -v ./...
 
 script:
     - go vet -x ./...
     - $HOME/gopath/bin/golint .
-    - go test -covermode=count -coverprofile=profile.cov -v .
+    - go test -v ./...
+    - go test -covermode=count -coverprofile=profile.cov .
 
 after_script:
     - $HOME/gopath/bin/goveralls -coverprofile=profile.cov -service=travis-ci
