+++ examples/arguments/arguments_example.go
