+++ examples/counted/counted_example.go
