+++ examples/git/branch/git_branch.go
@@ -1,4 +1,4 @@
-package main
+package git
 
 import (
 	"fmt"
