+++ examples/git/checkout/git_checkout.go
@@ -1,4 +1,4 @@
-package main
+package git
 
 import (
 	"fmt"
