+++ examples/git/clone/git_clone.go
@@ -1,4 +1,4 @@
-package main
+package git
 
 import (
 	"fmt"
