+++ examples/git/push/git_push.go
@@ -1,4 +1,4 @@
-package main
+package git
 
 import (
 	"fmt"
