+++ examples/git/remote/git_remote.go
@@ -1,4 +1,4 @@
-package main
+package git
 
 import (
 	"fmt"
