+++ examples/naval_fate/naval_fate.go
