+++ examples/options/options_example.go
