+++ examples/options_shortcut/options_shortcut_example.go
