+++ examples/quick/quick_example.go
