+++ examples/type_assert/type_assert_example.go
